requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.0
playwright==1.47.0
pandas==2.2.3
pypdf==5.1.0
python-multipart==0.0.9
matplotlib==3.9.2
numpy==2.1.1
python-dotenv==1.0.1